Refactor import statements in route.js files and update image paths in components
