--- src/models/userModel.js
@@ -1,31 +1,0 @@
-import mongoose from "mongoose";
-
-const userSchema = new mongoose.Schema(
-  {
-    name: {
-      type: String,
-    },
-    email: {
-      type: String,
-    },
-    blogs: {
-      type: [
-        {
-          type: mongoose.Schema.Types.ObjectId,
-          ref: "Blog",
-        },
-      ],
-      default: [],
-    },
-  },
-  { timestamps: true }
-);
-
-let User;
-try {
-  User = mongoose.model("User");
-} catch (error) {
-  User = mongoose.model("User", userSchema);
-}
-
-export default User;
